fix: Calcular CER T-10 dinámicamente desde serie histórica
- Actualizar /api/cer para calcular T-10 desde CER_serie.csv (local)
- Actualizar /api/cer-ultimo para leer último CER desde CSV (local)
- Actualizar función serverless cer.js para calcular T-10 desde API BCRA
- Agregar función serverless cer-ultimo.js para último CER
- Agregar redirect /api/cer-ultimo en netlify.toml
- TIR ahora se calculan con CER T-10 correcto
- UI muestra último CER publicado
